move postgres module
