+++ benchrunner-service/src/postgres/mod.rs
@@ -1,1 +1,2 @@
 pub mod metrics_dbstore;
+pub mod postgres_session;
